Add support for episode selection
Initial support for selecting indivual TV episodes has been added.
As there is no way to tell which episodes have been watched via Netflix,
I've also included a resume button to every show with multiple episode
to enable the user to continue watching where they left off. There are
a small number of one off shows that Netflix classes as a TV Programme
but have no individual episodes of their own. This is typically the case
with the "Best of SNL" specials. For the time being they appear as a
single episode within the show, ideally I'd like to reclass these one
off TV shows as a movie. I'll look into the best way to do this in the
future without adjusting the raw data from Netflix.