Fix search bug
It helps if I rememeber to encode the user input before attaching it
to a url. The search function should now be working correctly.

I've also updated the error handling for the JSON lookups so that
plex doesn't freak out when it receives a 404 response.

For future reference, the framework exception handler is called Ex.