Reduce the default resolution
The resolution has been reduced to 720p to help out with some transcoding
issues a few users are experiencing when viewing video on a different
machine to the Plex Media Server. I'd like to enable the users to change
the resolution of the webkit player in the preferences menu however I
don't believe this is possible at the present.